Fix a bug where a view is pinned to UITransitionView that is no longer shown
Since we are listening to the `didMoveToWindowSignal` we don't get notified when the view we're pinning to is removed from its window. Using the `windowSignal` fixes that problem but a transition view might still not be deallocated after it was removed from the window in which case we would try to pin our view to it but we will get a crash because the views are not in the same hierarchy (that's why the `isDescendant(of:)` check is needed). Additionally, we were keeping a strong reference in a block to the transition view which prevented if from being deallocated which is resolved with the added `[weak parent]`.