Add a text check before skipping label text and style update
Even though using `value` instead of `text` is recommended "to not accidentically break styling" using `text` on UILabel is very common and can accidentaly cause bugs hard to track down. This change makes it harder to do so. When using style with custom attributes it is possible to end up in a situation where the label is not showing any text even if its `value` says otherwise. One case like this is covered by `testSettingSameLabelValueAfterResettingText_styleWithCustomAttributes` unit test.