Remove wrong scrollview top pinning constant on iOS 11, leftover from iOS 10
The content insets behaviour on iOS 10 was weird, the insets of the scrollview by default included the nav bar adjustment because the scrollviews would go below the navigation bars but there was no good way to get only those insets without the custom set insets. I guess the iOS 10 pinning solution accomodates for that using the `frame.origin.y` as a constant for the pinned view constraint. That assumes the scrollview is always placed at the top of the screen. This PR doesn't address this logic that can be potentially causing bugs for some users.
On iOS  11 we use the frameLayoutGuide which takes into account the safe area and bars so using the originY of the scrollview as a constraint constant causes a gap between the pinned view and the navigation bar.